test: move tests into own directory
Keeps the snapshots separate from test files.
